requirements.txt
streamlit==1.31.0
   langchain==0.1.9
   langchain-openai==0.0.8
   langchain-google-genai==0.0.11
   langchain-anthropic==0.1.1
   openai==1.12.0
   google-generativeai>=0.4.1,<0.5.0
   anthropic==0.18.0
   shopify-python-api==12.6.0
   requests==2.31.0
